requirements.txt
asgiref==3.9.0
certifi==2025.7.9
cloudinary==1.44.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.1
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
